requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.8
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
djangorestframework-jwt==1.11.0

requests~=2.24.0